ai_reels_generator/pyproject.toml
[project]
name = "ai-reels-generator"
version = "0.1.0"
description = "CrewAI-based workflow for turning long-form videos into short vertical reels."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "crewai>=0.5.0",
  "pydantic>=2.10.6",
  "python-dotenv>=1.0.1",
  "openai>=1.68.0",
  "PyYAML>=6.0.2",
  "Flask>=3.1.0",
  "gunicorn>=23.0.0",
  "google-api-python-client>=2.164.0",
  "google-auth>=2.38.0",
  "google-auth-httplib2>=0.2.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.5",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
